Add new cmsdas catalogue and refactor slightly existing code
